backend/requirements.txt
Flask
SQLAlchemy
psycopg2
bcrypt
python-dotenv